Added functions to analyse actions
First attempt to try to understand action effects at a domain level
(find useful effects and remove the others)
Added functions to analyse which actions set a goal predicate and logic
to give rewards to plans that implement these actions.
Function to analyse which goals should be rewarded and when added.