Added precondition to metric actions detection
In order to get more cooperative solutions, it's necessary to detect
which actions allow the agents to execute metric dependent actions.
It's important to find which of this kind of actions are optimizers of
the metric (changing nav mode), and which are mandatory (undock). After
this distinction is made, metric cooperativeness rewards will be
introduced.